+++ models/order.go
@@ -6,14 +6,14 @@ import (
 )
 
 type Order struct {
-	OrderID         uint64    `json:"orderid" structs:"orderid"`
+	OrderID         uint32    `json:"orderid" structs:"orderid"`
 	OrderTime       time.Time `json:"ordertime" structs:"ordertime"`
 	RequesterName   string    `json:"requestername" structs:"requestername"`
 	AcceptorName    string    `json:"acceptorname" structs:"acceptorname"`
 	OrderedMealName string    `json:"mealname" structs:"mealname"`
-	RequesterId     uint64    `json:"requesterId" structs:"requesterId"`
-	AcceptorId      uint64    `json:"acceptorId" structs:"acceptorId"`
-	OrderedMealId   uint64    `json:"mealId" structs:"mealId"`
+	RequesterId     uint32    `json:"requesterId" structs:"requesterId"`
+	AcceptorId      uint32    `json:"acceptorId" structs:"acceptorId"`
+	OrderedMealId   uint32    `json:"mealId" structs:"mealId"`
 	IsDone          bool      `json:"isdone" structs:"isdone"`
 }
